Added data tracked
Now it tracks specific ranges of damage and compares damage to a critical amount to win a game with. Additionally, some logic hole that allowed reckless bushwhacker to sit in hand have been fixed along with moving most uses of Ghor-Clan Rampager to the end of the firstMain function.


TODO: 
		Check colors of mana before playing cards
		Comment the code